requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
feedparser>=6.0.11
urllib3>=2.0.0
